# docs/intel接口文档-中文版本-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Host Service",
        "description": "主机管理和WebSocket实时通信服务",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/host/file/upload": {
            "post": {
                "tags": [
                    "文件管理"
                ],
                "summary": "上传文件",
                "description": "上传文件到服务器指定目录，返回文件访问 URL",
                "operationId": "upload_file_api_v1_host_file_upload_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_file_api_v1_host_file_upload_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_FileUploadResponse_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/host/file/{filename}": {
            "get": {
                "tags": [
                    "文件管理"
                ],
                "summary": "获取文件（支持断点续传）",
                "description": "通过 `saved_filename` 获取上传的文件内容，默认返回完整文件。\n\n### Range 下载说明\n- 支持标准 HTTP Range 请求头，格式：`Range: bytes=start-end`\n- 例如：`Range: bytes=0-1048575` 可获取前 1MB，用于断点续传\n- 响应会携带 `Accept-Ranges: bytes`、`Content-Range`、`Content-Length`\n- 当 Range 合法时返回 `206 Partial Content`，否则返回 `416`\n- 未携带 Range 时返回完整文件（200 OK）",
                "operationId": "get_file_api_v1_host_file__filename__get",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "保存的文件名（由上传接口返回的 saved_filename）",
                            "title": "Filename"
                        },
                        "description": "保存的文件名（由上传接口返回的 saved_filename）"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回完整文件内容",
                        "content": {
                            "application/json": {
                                "schema": {}
                            },
                            "application/octet-stream": {}
                        }
                    },
                    "206": {
                        "description": "返回部分内容（断点续传），包含 Content-Range 头",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "文件不存在"
                    },
                    "416": {
                        "description": "Range 请求范围无效"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_file_api_v1_host_file_upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File",
                        "description": "要上传的文件"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_file_api_v1_host_file_upload_post"
            },
            "FileUploadResponse": {
                "properties": {
                    "file_id": {
                        "type": "string",
                        "title": "File Id",
                        "description": "文件唯一标识"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename",
                        "description": "原始文件名"
                    },
                    "saved_filename": {
                        "type": "string",
                        "title": "Saved Filename",
                        "description": "保存的文件名"
                    },
                    "file_url": {
                        "type": "string",
                        "title": "File Url",
                        "description": "文件访问 URL"
                    },
                    "file_size": {
                        "type": "integer",
                        "title": "File Size",
                        "description": "文件大小（字节）"
                    },
                    "content_type": {
                        "type": "string",
                        "title": "Content Type",
                        "description": "文件 MIME 类型"
                    },
                    "upload_time": {
                        "type": "string",
                        "title": "Upload Time",
                        "description": "上传时间"
                    }
                },
                "type": "object",
                "required": [
                    "file_id",
                    "filename",
                    "saved_filename",
                    "file_url",
                    "file_size",
                    "content_type",
                    "upload_time"
                ],
                "title": "FileUploadResponse",
                "description": "文件上传响应模式"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Result_FileUploadResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "响应码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "响应消息",
                        "default": "操作成功"
                    },
                    "data": {
                        "$ref": "#/components/schemas/FileUploadResponse",
                        "description": "响应数据"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "响应时间戳（ISO 8601 格式）"
                    },
                    "locale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Locale",
                        "description": "语言代码（用于多语言支持）"
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "Result[FileUploadResponse]"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}